# Conflicts:
#	.idea/deploymentTargetSelector.xml

+++ app/src/main/java/com/appmoviles/utadeliciasapp/ProductosFragmentos.kt
@@ -1,6 +1,6 @@
 package com.appmoviles.utadeliciasapp
 
-
+import android.app.AlertDialog
 import android.os.Bundle
 import androidx.fragment.app.Fragment
 import android.view.LayoutInflater
@@ -9,6 +9,7 @@ import android.view.ViewGroup
 import android.widget.Button
 import androidx.recyclerview.widget.LinearLayoutManager
 import androidx.recyclerview.widget.RecyclerView
+import com.google.firebase.auth.FirebaseAuth
 import com.google.firebase.firestore.FirebaseFirestore
 
 class ProductosFragmentos : Fragment(), ProductsAdapter.OnItemClickListener {
@@ -17,7 +18,6 @@ class ProductosFragmentos : Fragment(), ProductsAdapter.OnItemClickListener {
     private lateinit var btnAddProduct: Button
     private lateinit var recyclerView: RecyclerView
     private lateinit var adapter: ProductsAdapter
-    private lateinit var add: AddProductsFragment
 
     private val db = FirebaseFirestore.getInstance()
     private val productosColeccion = db.collection("Products")
@@ -32,7 +32,6 @@ class ProductosFragmentos : Fragment(), ProductsAdapter.OnItemClickListener {
     override fun onViewCreated(view: View, savedInstanceState: Bundle?) {
         super.onViewCreated(view, savedInstanceState)
 
-        add = AddProductsFragment()
         btnAddProduct = view.findViewById(R.id.btnaddProduct)
         recyclerView = view.findViewById(R.id.rvproducts)
         recyclerView.layoutManager = LinearLayoutManager(requireContext())
@@ -48,39 +47,57 @@ class ProductosFragmentos : Fragment(), ProductsAdapter.OnItemClickListener {
                 .addToBackStack(null)
                 .commit()
         }
-
-
-
     }
 
     private fun consultarColeccion() {
-        productosColeccion.get()
+        val userId = FirebaseAuth.getInstance().currentUser?.uid ?: ""
+        productosColeccion.whereEqualTo("userId", userId).get()
             .addOnSuccessListener { querySnapshot ->
                 val listaProductos = mutableListOf<Products>()
                 for (document in querySnapshot) {
                     val nombre = document.getString("Nombre")
                     val descripcion = document.getString("Descripción")
                     val imagen = document.getString("ImagenUrl") ?: ""
-                    //val precio = document.getString("precio")?.toInt()
                     val txtcantidad = document.getLong("Cantidad")?.toInt()
                     val ID = document.id
                     if (nombre != null && descripcion != null && txtcantidad != null) {
-                        val producto = Products(ID, nombre, descripcion, imagen, txtcantidad)//,precio)
+                        val producto = Products(ID, nombre, descripcion, imagen, txtcantidad)
                         listaProductos.add(producto)
                     }
                 }
                 adapter.setDatos(listaProductos)
             }
+            .addOnFailureListener { e ->
+                // Manejar el error en caso de fallo
+            }
     }
 
     override fun onDeleteItemClick(productId: String) {
-        productosColeccion.document(productId).delete()
-            .addOnSuccessListener {
-                consultarColeccion()  // Actualizar lista
+        // Diálogo de confirmación antes de eliminar
+        AlertDialog.Builder(context)
+            .setTitle("Eliminar Producto")
+            .setMessage("¿Estás seguro de que deseas eliminar este producto?")
+            .setPositiveButton("Sí") { dialog, which ->
+                productosColeccion.document(productId).delete()
+                    .addOnSuccessListener {
+                        consultarColeccion()  // Actualizar lista
+                        val transaction = parentFragmentManager.beginTransaction()
+                        transaction.replace(
+                            R.id.frame_layout,
+                            Producto_eliminado_exitosamente()
+                        ) // Asegúrate de usar el ID correcto de tu contenedor de fragmentos
+                        transaction.addToBackStack(null) // Opcional: para permitir volver al fragmento anterior
+                        transaction.commit()
+                    }
+                    .addOnFailureListener { e ->
+                        // Manejar el error en caso de fallo en la eliminación
+                    }
+
+                dialog.dismiss()
             }
-            .addOnFailureListener { e ->
-                // Manejar el error en caso de fallo en la eliminación
-            }
+            .setNegativeButton("No", null)
+            .show()
+
+
     }
 }
-
